Bug fixes & code optimization - Fixed issue with inability to use map-integrated chairs - Player and GM commands are now processed via thread executor - Players will no longer be locked out of their account if already logged-in - Repeating respawn task moved to its own class: FieldUpdateTask - FieldUpdateTask calls itself via thread executor - Removed RunningOperation enum (was a stupid idea) - Moved some data initialization to the channel server from server init
Commands are now executed asynchronously due to some commands being length and freezing the user socket thread.

Upon logging-in an account that has an existing player entity in-game, the existing user will be forced to log-out.

Refactoring tasks with long fixed intervals to scheduled threads that call themselves repeatedly in slightly shorter intervals

The Discord SQL connection is not established unless the inter listener is initialized